[Fix] 리스트 뷰 외부 컨테이너 스크롤 방지
